# Conflicts:
#	agrirouter-middleware-application/src/main/java/de/agrirouter/middleware/config/SpringSecurityConfiguration.java
#	agrirouter-middleware-business/src/main/java/de/agrirouter/middleware/business/TenantService.java
#	agrirouter-middleware-business/src/main/java/de/agrirouter/middleware/business/cache/messaging/MessageCache.java
#	agrirouter-middleware-controller/src/main/java/de/agrirouter/middleware/controller/dto/request/OnboardProcessRequest.java
#	agrirouter-middleware-controller/src/main/java/de/agrirouter/middleware/controller/secured/ApplicationController.java
#	agrirouter-middleware-controller/src/main/java/de/agrirouter/middleware/controller/secured/EndpointController.java
#	agrirouter-middleware-controller/src/main/java/de/agrirouter/middleware/controller/secured/NonTelemetryDataController.java
#	agrirouter-middleware-controller/src/main/java/de/agrirouter/middleware/controller/secured/TelemetryDataController.java
#	agrirouter-middleware-controller/src/main/java/de/agrirouter/middleware/controller/secured/TelemetryPlatformController.java
#	agrirouter-middleware-controller/src/main/java/de/agrirouter/middleware/controller/unsecured/maintenance/TenantMaintenanceController.java
#	agrirouter-middleware-domain/src/main/java/de/agrirouter/middleware/domain/ConnectionCriteria.java

+++ agrirouter-middleware-business/src/main/java/de/agrirouter/middleware/business/SearchNonTelemetryDataService.java
@@ -6,7 +6,7 @@ import de.agrirouter.middleware.api.errorhandling.error.ErrorMessageFactory;
 import de.agrirouter.middleware.business.dto.MessageStatistics;
 import de.agrirouter.middleware.business.parameters.SearchNonTelemetryDataParameters;
 import de.agrirouter.middleware.domain.ContentMessageMetadata;
-import de.agrirouter.middleware.persistence.ContentMessageRepository;
+import de.agrirouter.middleware.persistence.jpa.ContentMessageRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Service;
@@ -39,26 +39,32 @@ public class SearchNonTelemetryDataService {
      * @return -
      */
     public List<ContentMessageMetadata> search(SearchNonTelemetryDataParameters searchNonTelemetryDataParameters) {
-        final var endpoint = endpointService.findByExternalEndpointId(searchNonTelemetryDataParameters.getExternalEndpointId());
-        log.debug("Searching for existing messages, since the endpoint was found.");
-        log.trace("Filter criteria are >>> {}.", searchNonTelemetryDataParameters);
-        final List<ContentMessageMetadata> contentMessageMetadata;
-        if (null != searchNonTelemetryDataParameters.getTechnicalMessageTypes() && !searchNonTelemetryDataParameters.getTechnicalMessageTypes().isEmpty()) {
-            contentMessageMetadata = contentMessageRepository.findMetadata(endpoint.getAgrirouterEndpointId(),
-                    searchNonTelemetryDataParameters.getTechnicalMessageTypes().stream().map(ContentMessageType::getKey).toList(),
-                    List.of(ContentMessageType.ISO_11783_TIME_LOG.getKey(), ContentMessageType.ISO_11783_DEVICE_DESCRIPTION.getKey()),
-                    searchNonTelemetryDataParameters.getSendFrom(),
-                    searchNonTelemetryDataParameters.getSendTo());
-            log.debug("Found {} content messages in total.", contentMessageMetadata.size());
+        final var optionalEndpoint = endpointService.findByExternalEndpointId(searchNonTelemetryDataParameters.getExternalEndpointId());
+        if (optionalEndpoint.isPresent()) {
+            var endpoint = optionalEndpoint.get();
+            log.debug("Searching for existing messages, since the endpoint was found.");
+            log.trace("Filter criteria are >>> {}.", searchNonTelemetryDataParameters);
+            final List<ContentMessageMetadata> contentMessageMetadata;
+            if (null != searchNonTelemetryDataParameters.getTechnicalMessageTypes() && !searchNonTelemetryDataParameters.getTechnicalMessageTypes().isEmpty()) {
+                contentMessageMetadata = contentMessageRepository.findMetadata(endpoint.getAgrirouterEndpointId(),
+                        searchNonTelemetryDataParameters.getTechnicalMessageTypes().stream().map(ContentMessageType::getKey).toList(),
+                        List.of(ContentMessageType.ISO_11783_TIME_LOG.getKey(), ContentMessageType.ISO_11783_DEVICE_DESCRIPTION.getKey()),
+                        searchNonTelemetryDataParameters.getSendFrom(),
+                        searchNonTelemetryDataParameters.getSendTo());
+                log.debug("Found {} content messages in total.", contentMessageMetadata.size());
+            } else {
+                contentMessageMetadata = contentMessageRepository.findMetadata(endpoint.getAgrirouterEndpointId(),
+                        List.of(ContentMessageType.ISO_11783_TIME_LOG.getKey(), ContentMessageType.ISO_11783_DEVICE_DESCRIPTION.getKey()),
+                        searchNonTelemetryDataParameters.getSendFrom(),
+                        searchNonTelemetryDataParameters.getSendTo());
+            }
+            var flattenedContentMessageMetadata = flattenContentMessageMetadata(contentMessageMetadata);
+            log.debug("The {} content messages are flattened to {} 'real' messages.", contentMessageMetadata.size(), flattenedContentMessageMetadata);
+            return flattenedContentMessageMetadata;
         } else {
-            contentMessageMetadata = contentMessageRepository.findMetadata(endpoint.getAgrirouterEndpointId(),
-                    List.of(ContentMessageType.ISO_11783_TIME_LOG.getKey(), ContentMessageType.ISO_11783_DEVICE_DESCRIPTION.getKey()),
-                    searchNonTelemetryDataParameters.getSendFrom(),
-                    searchNonTelemetryDataParameters.getSendTo());
+            log.warn("The endpoint with the external ID '{}' was not found, therefore we cannot search for messages.", searchNonTelemetryDataParameters.getExternalEndpointId());
+            return Collections.emptyList();
         }
-        var flattenedContentMessageMetadata = flattenContentMessageMetadata(contentMessageMetadata);
-        log.debug("The {} content messages are flattened to {} 'real' messages.", contentMessageMetadata.size(), flattenedContentMessageMetadata);
-        return flattenedContentMessageMetadata;
     }
 
     private List<ContentMessageMetadata> flattenContentMessageMetadata(List<ContentMessageMetadata> contentMessageMetadata) {
@@ -98,19 +104,25 @@ public class SearchNonTelemetryDataService {
     }
 
     private byte[] download(String externalEndpointId, String messageId) {
-        final var endpoint = endpointService.findByExternalEndpointId(externalEndpointId);
-        final var optionalContentMessage = contentMessageRepository.findFirstByAgrirouterEndpointIdAndContentMessageMetadataMessageId(endpoint.getAgrirouterEndpointId(), messageId);
-        if (optionalContentMessage.isPresent()) {
-            final var contentMessage = optionalContentMessage.get();
-            if (contentMessage.getContentMessageMetadata().getTotalChunks() > 1) {
-                log.debug("Looks like we have multiple chunks for the content message. Assembling the message content first. There are {} chunks in total.", contentMessage.getContentMessageMetadata().getTotalChunks());
-                return assembleChunkedMessageContent(endpoint.getAgrirouterEndpointId(), contentMessage.getContentMessageMetadata().getChunkContextId());
+        final var optionalEndpoint = endpointService.findByExternalEndpointId(externalEndpointId);
+        if (optionalEndpoint.isPresent()) {
+            var endpoint = optionalEndpoint.get();
+            final var optionalContentMessage = contentMessageRepository.findFirstByAgrirouterEndpointIdAndContentMessageMetadataMessageId(endpoint.getAgrirouterEndpointId(), messageId);
+            if (optionalContentMessage.isPresent()) {
+                final var contentMessage = optionalContentMessage.get();
+                if (contentMessage.getContentMessageMetadata().getTotalChunks() > 1) {
+                    log.debug("Looks like we have multiple chunks for the content message. Assembling the message content first. There are {} chunks in total.", contentMessage.getContentMessageMetadata().getTotalChunks());
+                    return assembleChunkedMessageContent(endpoint.getAgrirouterEndpointId(), contentMessage.getContentMessageMetadata().getChunkContextId());
+                } else {
+                    log.debug("This is a single message, therefore returning the content 'as it is'.");
+                    return Base64.getDecoder().decode(contentMessage.getMessageContent());
+                }
             } else {
-                log.debug("This is a single message, therefore returning the content 'as it is'.");
-                return Base64.getDecoder().decode(contentMessage.getMessageContent());
+                throw new BusinessException(ErrorMessageFactory.couldNotFindContentMessage());
             }
         } else {
-            throw new BusinessException(ErrorMessageFactory.couldNotFindContentMessage());
+            log.warn("The endpoint with the external ID '{}' was not found, therefore we cannot search for messages.", externalEndpointId);
+            return new byte[0];
         }
     }
 
@@ -146,17 +158,22 @@ public class SearchNonTelemetryDataService {
     public MessageStatistics getMessageStatistics(String externalEndpointId) {
         var messageStatistics = new MessageStatistics();
         messageStatistics.setExternalEndpointId(externalEndpointId);
-        final var endpoint = endpointService.findByExternalEndpointId(externalEndpointId);
-        var messageCountForTechnicalMessageTypes = contentMessageRepository.countMessagesGroupedByTechnicalMessageType(endpoint.getAgrirouterEndpointId());
-        messageCountForTechnicalMessageTypes.forEach(messageCountForTechnicalMessageType -> {
-                    log.debug("Found {} messages for technical message type {} for the sender {}.",
-                            messageCountForTechnicalMessageType.getNumberOfMessages(),
-                            messageCountForTechnicalMessageType.getTechnicalMessageType(),
-                            messageCountForTechnicalMessageType.getSenderId());
-                    messageStatistics.addMessageStatisticEntry(messageCountForTechnicalMessageType.getSenderId(),
-                            new MessageStatistics.MessageStatistic.Entry(messageCountForTechnicalMessageType.getTechnicalMessageType(), messageCountForTechnicalMessageType.getNumberOfMessages()));
-                }
-        );
+        final var optionalEndpoint = endpointService.findByExternalEndpointId(externalEndpointId);
+        if (optionalEndpoint.isPresent()) {
+            var endpoint = optionalEndpoint.get();
+            var messageCountForTechnicalMessageTypes = contentMessageRepository.countMessagesGroupedByTechnicalMessageType(endpoint.getAgrirouterEndpointId());
+            messageCountForTechnicalMessageTypes.forEach(messageCountForTechnicalMessageType -> {
+                        log.debug("Found {} messages for technical message type {} for the sender {}.",
+                                messageCountForTechnicalMessageType.getNumberOfMessages(),
+                                messageCountForTechnicalMessageType.getTechnicalMessageType(),
+                                messageCountForTechnicalMessageType.getSenderId());
+                        messageStatistics.addMessageStatisticEntry(messageCountForTechnicalMessageType.getSenderId(),
+                                new MessageStatistics.MessageStatistic.Entry(messageCountForTechnicalMessageType.getTechnicalMessageType(), messageCountForTechnicalMessageType.getNumberOfMessages()));
+                    }
+            );
+        } else {
+            log.warn("The endpoint with the external ID '{}' was not found, therefore we cannot search for messages.", externalEndpointId);
+        }
         return messageStatistics;
     }
 
@@ -168,20 +185,25 @@ public class SearchNonTelemetryDataService {
      */
     @Transactional
     public void delete(String externalEndpointId, String messageId) {
-        final var endpoint = endpointService.findByExternalEndpointId(externalEndpointId);
-        final var optionalContentMessage = contentMessageRepository.findFirstByAgrirouterEndpointIdAndContentMessageMetadataMessageId(endpoint.getAgrirouterEndpointId(), messageId);
-        if (optionalContentMessage.isPresent()) {
-            final var contentMessage = optionalContentMessage.get();
-            if (contentMessage.getContentMessageMetadata().getTotalChunks() > 1) {
-                log.debug("Looks like we have multiple chunks for the content message. Assembling the message content first. There are {} chunks in total.", contentMessage.getContentMessageMetadata().getTotalChunks());
-                deleteChunkedMessageContent(endpoint.getAgrirouterEndpointId(), contentMessage.getContentMessageMetadata().getChunkContextId());
+        final var optionalEndpoint = endpointService.findByExternalEndpointId(externalEndpointId);
+        if (optionalEndpoint.isPresent()) {
+            var endpoint = optionalEndpoint.get();
+            final var optionalContentMessage = contentMessageRepository.findFirstByAgrirouterEndpointIdAndContentMessageMetadataMessageId(endpoint.getAgrirouterEndpointId(), messageId);
+            if (optionalContentMessage.isPresent()) {
+                final var contentMessage = optionalContentMessage.get();
+                if (contentMessage.getContentMessageMetadata().getTotalChunks() > 1) {
+                    log.debug("Looks like we have multiple chunks for the content message. Assembling the message content first. There are {} chunks in total.", contentMessage.getContentMessageMetadata().getTotalChunks());
+                    deleteChunkedMessageContent(endpoint.getAgrirouterEndpointId(), contentMessage.getContentMessageMetadata().getChunkContextId());
+                } else {
+                    log.debug("This is a single message, therefore nothing else to do.");
+                    var i = contentMessageRepository.deleteByAgrirouterEndpointIdAndContentMessageMetadataMessageId(endpoint.getAgrirouterEndpointId(), messageId);
+                    log.debug("Deleted {} content message, no chunks were harmed.", i);
+                }
             } else {
-                log.debug("This is a single message, therefore nothing else to do.");
-                var i = contentMessageRepository.deleteByAgrirouterEndpointIdAndContentMessageMetadataMessageId(endpoint.getAgrirouterEndpointId(), messageId);
-                log.debug("Deleted {} content message, no chunks were harmed.", i);
+                throw new BusinessException(ErrorMessageFactory.couldNotFindContentMessage());
             }
         } else {
-            throw new BusinessException(ErrorMessageFactory.couldNotFindContentMessage());
+            log.warn("The endpoint with the external ID '{}' was not found, therefore we cannot search for messages.", externalEndpointId);
         }
     }
